+++ Tving_CloneProject/Tving_CloneProject/Scene/LoginViewController.swift
@@ -11,16 +11,12 @@ import SnapKit
 import Then
 
 class LoginViewController: UIViewController {
-
+    
     // MARK: - UI Properties
     
     private let loginLabel = UILabel()
     
-    private lazy var idTextField = UITextField()
-    
-    private lazy var pwTextField = UITextField()
-    
-    private let loginButton = UIButton()
+    private let loginView = LoginView()
     
     private let findIdLabel = UILabel()
     
@@ -32,16 +28,6 @@ class LoginViewController: UIViewController {
     
     private let createNicknameLabel = UILabel()
     
-    private let idButtonBox = UIView(frame: CGRect(x: 0, y: 0, width: 40, height: 20))
-    
-    private let pwButtonBox = UIView()
-    
-    private lazy var idClearButton = UIButton(frame: CGRect(x: 0, y: 0, width: 20, height: 20))
-    
-    private lazy var pwClearButton = UIButton(frame: CGRect(x: 0, y: 0, width: 20, height: 20))
-    
-    private lazy var maskButton =  UIButton(frame: CGRect(x: 0, y: 0, width: 20, height: 20))
-    
     
     // MARK: - Properties
     
@@ -52,17 +38,18 @@ class LoginViewController: UIViewController {
     var isActivate: Bool = false
     
     var nickname: String = ""
-
+    
+    
     // MARK: - Life Cycles
     
     override func viewDidLoad() {
         super.viewDidLoad()
-    
+        
         setHierarchy()
         setLayout()
         setStyle()
     }
-
+    
 }
 
 
@@ -71,10 +58,8 @@ class LoginViewController: UIViewController {
 private extension LoginViewController {
     
     func setHierarchy() {
-        self.view.addSubviews(loginLabel, 
-                              idTextField, 
-                              pwTextField,
-                              loginButton,
+        self.view.addSubviews(loginLabel,
+                              loginView,
                               findIdLabel,
                               findPwLabel,
                               divider,
@@ -89,42 +74,28 @@ private extension LoginViewController {
             $0.centerX.equalToSuperview()
         }
         
-        idTextField.snp.makeConstraints {
+        loginView.snp.makeConstraints {
             $0.top.equalTo(loginLabel.snp.bottom).offset(30)
             $0.centerX.equalToSuperview()
             $0.width.equalTo(width * 335)
-            $0.height.equalTo(height * 52)
-        }
-        
-        pwTextField.snp.makeConstraints {
-            $0.top.equalTo(idTextField.snp.bottom).offset(7)
-            $0.centerX.equalToSuperview()
-            $0.width.equalTo(width * 335)
-            $0.height.equalTo(height * 52)
-        }
-        
-        loginButton.snp.makeConstraints {
-            $0.top.equalTo(pwTextField.snp.bottom).offset(20)
-            $0.centerX.equalToSuperview()
-            $0.width.equalTo(width * 335)
-            $0.height.equalTo(height * 52)
+            $0.height.equalTo(height * 186)
         }
         
         findIdLabel.snp.makeConstraints {
-            $0.top.equalTo(loginButton.snp.bottom).offset(30)
+            $0.top.equalTo(loginView.snp.bottom).offset(30)
             $0.leading.equalToSuperview().inset(85)
             $0.width.equalTo(width * 65)
         }
         
         divider.snp.makeConstraints {
-            $0.top.equalTo(loginButton.snp.bottom).offset(30)
+            $0.top.equalTo(loginView.snp.bottom).offset(30)
             $0.centerX.equalToSuperview()
             $0.width.equalTo(1)
             $0.height.equalTo(12)
         }
         
         findPwLabel.snp.makeConstraints {
-            $0.top.equalTo(loginButton.snp.bottom).offset(30)
+            $0.top.equalTo(loginView.snp.bottom).offset(30)
             $0.leading.equalTo(divider.snp.trailing).offset(35)
             $0.width.equalTo(width * 75)
         }
@@ -139,10 +110,11 @@ private extension LoginViewController {
             $0.leading.equalTo(findPwLabel)
             $0.width.equalTo(130)
         }
-
+        
     }
     
     func setStyle() {
+        
         self.view.backgroundColor = UIColor(resource: .black)
         
         loginLabel.do {
@@ -151,71 +123,8 @@ private extension LoginViewController {
             $0.textColor = UIColor(resource: .grey1)
         }
         
-        idTextField.do {
-            $0.setTextField(forBackgroundColor: UIColor(resource: .grey4),
-                            forBorderColor: UIColor(resource: .grey4),
-                            forBorderWidth: 0,
-                            forCornerRadius: 3)
-            $0.setPlaceholder(placeholder: "아이디",
-                              fontColor: UIColor(resource: .grey2),
-                              font: UIFont.pretendard(.subhead3))
-            $0.setLeftPadding(amount: 22)
-            $0.textColor = UIColor(resource: .grey2)
-            
-            idButtonBox.addSubview(idClearButton)
-            idClearButton.snp.makeConstraints {
-                $0.top.leading.height.equalToSuperview()
-                $0.trailing.equalToSuperview().inset(20)
-            }
-            $0.rightView = idButtonBox
-            $0.rightViewMode = .whileEditing
-            
+        loginView.do {
             $0.delegate = self
-            $0.addTarget(self, action: #selector(textFieldChange), for: .editingChanged)
-        }
-        
-        pwTextField.do {
-            $0.setTextField(forBackgroundColor: UIColor(resource: .grey4),
-                            forBorderColor: UIColor(resource: .grey4),
-                            forBorderWidth: 0,
-                            forCornerRadius: 3)
-            $0.setPlaceholder(placeholder: "비밀번호",
-                              fontColor: UIColor(resource: .grey2),
-                              font: UIFont.pretendard(.subhead3))
-            $0.setLeftPadding(amount: 22)
-            $0.textColor = UIColor(resource: .grey2)
-            $0.isSecureTextEntry = true
-            
-            pwButtonBox.addSubviews(pwClearButton, maskButton)
-            
-            pwButtonBox.snp.makeConstraints {
-                $0.width.equalTo(80)
-                $0.height.equalTo(20)
-            }
-            maskButton.snp.makeConstraints {
-                $0.top.height.equalToSuperview()
-                $0.trailing.equalToSuperview().inset(20)
-            }
-            
-            pwClearButton.snp.makeConstraints {
-                $0.top.height.equalToSuperview()
-                $0.trailing.equalToSuperview().inset(56)
-            }
-            
-            $0.rightView = pwButtonBox
-            $0.rightViewMode = .whileEditing
-            
-            $0.delegate = self
-            $0.addTarget(self, action: #selector(textFieldChange), for: .editingChanged)
-        }
-        
-        loginButton.do {
-            $0.setTitle("로그인하기", for: .normal)
-            $0.setTitleColor(UIColor(resource: .grey2), for: .normal)
-            $0.layer.cornerRadius = 3
-            $0.layer.borderWidth = 1
-            $0.layer.borderColor = UIColor(resource: .grey4).cgColor
-            $0.addTarget(self, action: #selector(pushToWelcomeVC), for: .touchUpInside)
         }
         
         findIdLabel.do {
@@ -251,87 +160,6 @@ private extension LoginViewController {
             $0.addGestureRecognizer(gesture)
         }
         
-        idButtonBox.do {
-            $0.backgroundColor = UIColor.clear
-        }
-        
-        idClearButton.do {
-            $0.setImage(UIImage(resource: .clear), for: .normal)
-            $0.tag = 0
-            $0.addTarget(self, action: #selector(clearButtonTapped), for: .touchUpInside)
-        }
-        
-        pwButtonBox.do {
-            $0.backgroundColor = UIColor.clear
-        }
-        
-        pwClearButton.do {
-            $0.setImage(UIImage(resource: .clear), for: .normal)
-            $0.tag = 1
-            $0.addTarget(self, action: #selector(clearButtonTapped), for: .touchUpInside)
-        }
-        
-        maskButton.do {
-            $0.setImage(UIImage(systemName: "eye.slash.fill"), for: .normal)
-            $0.tintColor = UIColor(resource: .grey3)
-            $0.addTarget(self, action: #selector(maskButtonTapped), for: .touchUpInside)
-        }
-        
-    }
-    
-    @objc
-    func textFieldChange() {
-        let id = self.idTextField.text ?? ""
-        let pw = self.pwTextField.text ?? ""
-        
-        if !id.isEmpty {
-            idClearButton.isHidden = false
-        } else {
-            pwClearButton.isHidden = false
-            maskButton.isHidden = false
-        }
-        
-        if !id.isEmpty && !pw.isEmpty {
-            loginButton.backgroundColor = UIColor(resource: .red)
-            loginButton.setTitleColor(UIColor(resource: .white), for: .normal)
-            loginButton.layer.borderWidth = 0
-            loginButton.isEnabled = true
-            isActivate = true
-        } else {
-            loginButton.backgroundColor = UIColor(resource: .black)
-            loginButton.setTitleColor(UIColor(resource: .grey2), for: .normal)
-            loginButton.layer.borderWidth = 1
-            loginButton.isEnabled = false
-            isActivate = false
-        }
-    }
-    
-    @objc
-    func maskButtonTapped() {
-        self.pwTextField.isSecureTextEntry = !self.pwTextField.isSecureTextEntry
-    }
-    
-    @objc
-    func clearButtonTapped(_ sender: UIButton) {
-        if sender.tag == 0 {
-            self.idTextField.text = ""
-        } else {
-            self.pwTextField.text = ""
-        }
-        loginButton.backgroundColor = UIColor(resource: .black)
-        loginButton.setTitleColor(UIColor(resource: .grey2), for: .normal)
-        loginButton.layer.borderWidth = 1
-        loginButton.isEnabled = false
-        isActivate = false
-    }
-    
-    @objc
-    func pushToWelcomeVC() {
-        if isActivate {
-            let welcomeVC = WelcomeViewController()
-            welcomeVC.userInfo = nickname.isEmpty ? self.idTextField.text : nickname
-            self.navigationController?.pushViewController(welcomeVC, animated: true)
-        }
     }
     
     @objc
@@ -346,35 +174,18 @@ private extension LoginViewController {
 
 // MARK: - Delegates
 
-extension LoginViewController: UITextFieldDelegate {
-    
-    func textFieldDidBeginEditing (_ textField: UITextField) {
-        
-        if textField.placeholder == "아이디" {
-            self.idTextField.layer.borderWidth = 1
-            self.idTextField.layer.borderColor = UIColor(resource: .grey2).cgColor
-        } else {
-            self.pwTextField.layer.borderWidth = 1
-            self.pwTextField.layer.borderColor = UIColor(resource: .grey2).cgColor
-        }
-        
-    }
-    
-    func textFieldDidEndEditing(_ textField: UITextField) {
-        
-        if textField.placeholder == "아이디" {
-            self.idTextField.layer.borderWidth = 0
-            self.idTextField.layer.borderColor = UIColor(resource: .grey4).cgColor
-        } else {
-            self.pwTextField.layer.borderWidth = 0
-            self.pwTextField.layer.borderColor = UIColor(resource: .grey4).cgColor
-        }
-    }
-}
-
 extension LoginViewController: CreateNicknameVCDelegate {
     
     func saveUserNickname(nickname: String) {
         self.nickname = nickname
+    }
+}
+
+extension LoginViewController: LoginViewDelegate {
+    
+    func pushToWelcomeVC(id: String) {
+        let welcomeVC = WelcomeViewController()
+        welcomeVC.userInfo = nickname.isEmpty ? id : nickname
+        self.navigationController?.pushViewController(welcomeVC, animated: true)
     }
 }
